feat: 랜덤 값에 따른 자동차 전진 기능 구현
