+++ src/main/java/racingcar/domain/MoveStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.domain;
+
+public interface MoveStrategy {
+    boolean isMovable(int value);
+}
