+++ src/main/java/racingcar/domain/RandomStrategy.java
@@ -1,0 +1,8 @@
+package racingcar.domain;
+
+public class RandomStrategy implements MoveStrategy {
+    @Override
+    public boolean isMovable(int value) {
+        return 4 <= value;
+    }
+}
